requirements.txt
kivy
https://github.com/kivymd/KivyMD/archive/master.zip
signalslot
kivy_garden.graph
sqlalchemy==1.3.22
py_ballisticcalc==1.0.12
a7p==0.0.4
pyinstaller==5.6.2